requirements.txt
tweepy>=4.14.0
requests>=2.31.0
feedparser>=6.0.11
jinja2>=3.1.4
python-dotenv>=1.0.1
pytest>=8.0.0
pytest-mock>=3.14.0
